% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_files.R
\name{get_idx}
\alias{get_idx}
\title{get_idx}
\usage{
get_idx(table, col, pattern)
}
\arguments{
\item{table}{data.frame, Input dataframe.}

\item{col}{character, colomn to be searched.}

\item{pattern}{character, pattern to be searched in col values.}
}
\description{
returns the index of values matching a pattern in a given col
for a given table.
}
