% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_files.R
\name{import_lipid_data}
\alias{import_lipid_data}
\title{Import lipid data}
\usage{
import_lipid_data(
  data,
  blank_thr = 2,
  blank_thr_nr_of_samples = 0.8,
  group_thr_nr_of_samples = 0.7,
  norm_to_lipid = TRUE
)
}
\arguments{
\item{data}{data.frame, containing all data.}

\item{blank_thr}{integer, number of blank threshold.}

\item{blank_thr_nr_of_samples}{number, proportion of samples above blank
threshold.}

\item{group_thr_nr_of_samples}{number, proportion of samples from one group
above blank threshold.}

\item{norm_to_lipid}{boolean, total area normalization.}
}
\value{
list, with the following entries:
   data_class_norm: data.frame normalized lipid class
   data_tot_lip_norm: data.frame
   data_class_norm_melt: long format of data_class_norm
   class_data_tot_lip_norm: data.frame
   class_data_tot_lip_norm_melt: long format of class_data_tot_lipid_norm
   meta: data.frame with meta data
}
\description{
Import all lipid data.
}
\details{
files

The structure is slightly changed. Only the first sheet is read. The
    first column should be sampleID which contains several fields separated
    by '__'.
}
\author{
Yassene Mohammed

Rico Derks
}
